refactor: move src/Resume.tsx to src/components/resume.tsx
